test(fixture-helper): fix fixture dependencies [DIRTY HACK]
let androidModules = `node_modules/cordova-android/node_modules/`
let fixtureModules = `<APP_FIXTURE_PATH>/node_modules/`

- A: To make the `androidApp` fixture work, we link
  `fixtureModules` to our `node_modules/`
- B: `PlatformApi.createPlatform` copies `androidModules` to `fixtureModules` if present
- C: Previously there was no `androidModules` but with the latest version of cordova-android there is a `androidModules`
- D: Linking fails because `fixtureModules` already exists

The dirty hack employed here is to link to our `node_modules/` from the
parent dir of the fixture target directory. It works with the current
tests but this could easily lead to hard-to-diagnose bugs in the future.

The proper fix would be to fix `cordova-android` to stop copying
platform scripts to the platform project on creation and only require
`cordova-android` from the latter. This would alleviate the need for B,
make C irrelevant and fix D.